Restructured computation thread
